requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-jose
passlib[bcrypt]
pytz
alembic==1.15.2
nudenet
